pyproject.toml
[project]
name = "smart_mcp_server"
version = "0.1.0"
description = "The MCP server that is smart to insert and search different tools for you."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastmcp>=2.11.0",
    "ipykernel>=6.30.1",
    "langchain>=0.3.27",
    "langchain-community>=0.3.27",
    "langchain-mcp-adapters>=0.1.9",
    "langchain-openai>=0.3.32",
    "langgraph>=0.6.6",
    "python-dotenv>=1.1.1",
]
